+++ src/pages/homePage/homePage.js
@@ -1,8 +1,7 @@
 import React, { useState } from 'react';
 import { LinearGradient } from 'expo-linear-gradient';
-import { View, Text, StyleSheet, Dimensions, StatusBar } from 'react-native';
-import { customTheme } from '../../styles/customTheme.js';
-
+import { View, Text, StatusBar } from 'react-native';
+import HomeStyles from './homeStyles.js';
 
 import HomeBg from '../../assets/images/svg_js/home_bg.js';
 import LogoBr from '../../components/logo/logo_br.js';
@@ -21,11 +20,6 @@ import ChatInac24 from '../../components/icon_24/chat_inac_24.js';
 import HomeLine from '../../components/homeCompo/HomeLine.js';
 import HomecardBackBtn from '../../components/homeCompo/HomecardBackBtn.js';
 
-const screenWidth = Dimensions.get('window').width;
-const screenHeight = Dimensions.get('window').height;
-
-const { fontHead18, fontSub16, fontCaption } = customTheme;
-
 const HomePage = ({introduion = "adipiscing varius eu sit nulla, luctus tincidunt ex at ullamcorper cursus odio laoreet placerat.",
     name = "Amyyheart",
     country="France",
@@ -39,44 +33,44 @@ const HomePage = ({introduion = "adipiscing varius eu sit nulla, luctus tincidun
   };
 
   return (
-    <View style={styles.container}>
-      <LinearGradient style={[styles.linearGradient, { paddingTop: statusBarHeight }]} colors={['#0029F4', '#6199C1', '#6199C1']}>
-        <HomeBg style={styles.svg} preserveAspectRatio='none'/>
+    <View style={HomeStyles.container}>
+      <LinearGradient style={[HomeStyles.linearGradient, { paddingTop: statusBarHeight }]} colors={['#0029F4', '#6199C1', '#6199C1']}>
+        <HomeBg style={HomeStyles.svg} preserveAspectRatio='none'/>
         
-        <View style={styles.topContainer}>
-          <View style={styles.logo}>
+        <View style={HomeStyles.topContainer}>
+          <View style={HomeStyles.logo}>
             <LogoBr />
           </View>
-          <View style={styles.notify}>
+          <View style={HomeStyles.notify}>
             <Notify />
           </View>
         </View>
 
 
-        <View style={styles.textConnectWithContainer}>
-          <Text style={styles.TXconnect} >커넥트</Text>
-          <Text style={styles.TXwithnewfriend} >새로운 친구와 함께해요!</Text>
+        <View style={HomeStyles.textConnectWithContainer}>
+          <Text style={HomeStyles.TXconnect} >커넥트</Text>
+          <Text style={HomeStyles.TXwithnewfriend} >새로운 친구와 함께해요!</Text>
         </View>
 
 
         {showNewCard ? (
-          <View style={styles.homecardContainer}>
+          <View style={HomeStyles.homecardContainer}>
             <HomeArrow style={{transform: [{ scaleX: -1 }] }}/>
-            <View style={[styles.homecard, { alignItems: 'center' }]}>
+            <View style={[HomeStyles.homecard, { alignItems: 'center' }]}>
               <Homecard />
-              <View style={styles.homecardDifeF}>
+              <View style={HomeStyles.homecardDifeF}>
                 <HomecardDifeF />
               </View>
-              <View style={styles.homecardBack}>
+              <View style={HomeStyles.homecardBack}>
                 <HomeProfile profile={require('../../assets/images/test_img/test_haedam.jpg')}/>
-                <Text style={styles.viewProfile}>프로필 상세보기</Text>
-                <View style={styles.addFriendOk}>
-                  <Text style={styles.TXname}>{name}</Text>
-                  <Text style={styles.myinfo}>에게 친구신청하시겠습니까?</Text>
+                <Text style={HomeStyles.viewProfile}>프로필 상세보기</Text>
+                <View style={HomeStyles.addFriendOk}>
+                  <Text style={HomeStyles.TXname}>{name}</Text>
+                  <Text style={HomeStyles.myinfo}>에게 친구신청하시겠습니까?</Text>
                 </View>
                 
               </View>
-              <View style={styles.homecardBackBtn}>
+              <View style={HomeStyles.homecardBackBtn}>
                   <HomecardBackBtn btnText="아니오" onPress={() => setShowNewCard(false)} />
                   <HomecardBackBtn btnText="신청하기"/>
                 </View>
@@ -84,32 +78,32 @@ const HomePage = ({introduion = "adipiscing varius eu sit nulla, luctus tincidun
             <HomeArrow />
           </View>
         ) : (
-          <View style={styles.homecardContainer}>
+          <View style={HomeStyles.homecardContainer}>
             <HomeArrow style={{transform: [{ scaleX: -1 }] }}/>
-            <View style={styles.homecard}>
+            <View style={HomeStyles.homecard}>
               <Homecard />
-              <View style={styles.homecardDifeF}>
+              <View style={HomeStyles.homecardDifeF}>
                 <HomecardDifeF />
               </View>
-            <View style={styles.homeProfile}>
+            <View style={HomeStyles.homeProfile}>
               <HomeProfile profile={require('../../assets/images/test_img/test_haedam.jpg')}/>
-              <View style={styles.tagContainer}>
+              <View style={HomeStyles.tagContainer}>
                 <HomeTag text="Enfp"/>
                 <HomeTag text="Sports"/>
                 <HomeTag text="Drawing"/>
               </View>
-              <Text style={styles.introdution}>{introduion}</Text>
-              <View style={styles.myinfoContainer}>
-                <Text style={[styles.TXname, { fontWeight: 'NotoSansCJKkr-Medium' }]}>{name}</Text>
-                <Text style={styles.myinfo}> | {country} | {age}</Text>
+              <Text style={HomeStyles.introdution}>{introduion}</Text>
+              <View style={HomeStyles.myinfoContainer}>
+                <Text style={[HomeStyles.TXname, { fontWeight: 'NotoSansCJKkr-Medium' }]}>{name}</Text>
+                <Text style={HomeStyles.myinfo}> | {country} | {age}</Text>
               </View>
               </View>
-              <View style={styles.connectIconContainer}>
-                  <HeartInac24 style={styles.connectIcon}/>
-                  <HomeLine style={styles.connectIcon}/>
-                  <AddFriendInac24 style={styles.connectIcon} onPress={handleAddFriendPress}/>
-                  <HomeLine style={styles.connectIcon}/>
-                  <ChatInac24 style={styles.connectIcon}/>
+              <View style={HomeStyles.connectIconContainer}>
+                  <HeartInac24 style={HomeStyles.connectIcon}/>
+                  <HomeLine style={HomeStyles.connectIcon}/>
+                  <AddFriendInac24 style={HomeStyles.connectIcon} onPress={handleAddFriendPress}/>
+                  <HomeLine style={HomeStyles.connectIcon}/>
+                  <ChatInac24 style={HomeStyles.connectIcon}/>
               </View>
             </View>
           <HomeArrow />
@@ -117,177 +111,19 @@ const HomePage = ({introduion = "adipiscing varius eu sit nulla, luctus tincidun
         )}
 
 
-        <View style={styles.homeSchEv}>
+        <View style={HomeStyles.homeSchEv}>
           <HomeSchEv />
-          <Text style={styles.TXhomeSch}>학교정보</Text>
-          <HomeSchoolInfo style={styles.homeSchIcon}/>
-          <View style={styles.homeEv}>
+          <Text style={HomeStyles.TXhomeSch}>학교정보</Text>
+          <HomeSchoolInfo style={HomeStyles.homeSchIcon}/>
+          <View style={HomeStyles.homeEv}>
             <HomeSchEv />
-            <Text style={styles.TXhomeEv}>이벤트</Text>
-            <HomeEvent style={styles.homeEvIcon}/>
+            <Text style={HomeStyles.TXhomeEv}>이벤트</Text>
+            <HomeEvent style={HomeStyles.homeEvIcon}/>
           </View>
         </View>
       </LinearGradient>
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    flexDirection: 'row',
-  },
-  linearGradient: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  svg: {
-    position: 'absolute',
-    width: screenWidth,
-    height: screenHeight,
-  },
-  topContainer: {
-    flexDirection: 'row',
-  },
-  logo: {
-    flex: 1,
-    marginTop: 1,
-    marginLeft: -8
-  },
-  notify: {
-    marginTop: 7,
-    marginRight: 25,
-  },
-  textConnectWithContainer: {
-    flexDirection: 'row',
-    alignItems: 'flex-end',
-    marginTop: 25,
-    marginRight: 83,
-  },
-  TXconnect: {
-    ...fontHead18,
-    color: customTheme.bgList,
-    marginRight: 4,
-  },
-  TXwithnewfriend: {
-    ...fontCaption,
-    color: customTheme.bgList,
-  },
-  homecardContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  homecard: {
-    position: 'relative',
-    top: 8,
-    shadowColor: '#3C454E4A',
-    shadowOffset: { width: 0, height: 3 },
-    shadowOpacity: 0.29,
-    shadowRadius: 6,
-  },
-  homecardDifeF: {
-    position: 'absolute',
-    top: 70,
-  },
-  homeProfile: {
-    position: 'absolute',
-    left: 20,
-    top: 20,
-  },
-  tagContainer: {
-    flexDirection: 'row',
-    marginTop: 12,
-    marginBottom: 6,
-  },
-  introdution: {
-    ...fontCaption,
-    width: 200,
-    height: 51,
-    marginTop: 6,
-    marginBottom: 6,
-  },
-  myinfoContainer: {
-    flexDirection: 'row',
-    // marginTop: 6,
-  },
-  TXname: {
-    fontSize: 12,
-    lineHeight: 16,
-    fontFamily: 'NotoSansCJKkr-Bold',
-  },
-  myinfo: {
-    ...fontCaption,
-  },
-  connectIconContainer: {
-    position: 'absolute',
-    left: 10,
-    bottom: 28,
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  connectIcon: {
-    marginHorizontal: 28,
-  },
-  homeSchEv: {
-    flexDirection: 'row',
-    marginTop: 28,
-    shadowColor: '#3C454E4A',
-    shadowOffset: { width: 0, height: 3 },
-    shadowOpacity: 0.29,
-    shadowRadius: 6,
-  },
-  TXhomeSch: {
-    ...fontSub16,
-    position: 'absolute',
-    top: 12,
-    left: 12,
-  },
-  homeSchIcon: {
-    position: 'absolute',
-    top: 35,
-    left: 40,
-  },
-  homeEv: {
-    position: 'relative',
-    marginLeft: 21,
-  },
-  TXhomeEv: {
-    ...fontSub16,
-    position: 'absolute',
-    top: 12,
-    left: 12,
-  },
-  homeEvIcon: {
-    position: 'absolute',
-    top: 48,
-    left: 47,
-  },
-
-
-  homecardBack: {
-    position: 'absolute',
-    justifyContent: 'center',
-    alignItems: 'center',
-    top: 30,
-  },
-  viewProfile: {
-    ...fontCaption,
-    color: customTheme.textSecondary,
-    marginTop: 20,
-    textDecorationLine: 'underline'
-  },
-  addFriendOk: {
-    flexDirection: 'row',
-    marginTop: 33,
-  },
-  homecardBackBtn: {
-    position: 'absolute',
-    flexDirection: 'row',
-    bottom: 20,
-  },
-  
-});
 
 export default HomePage;
